refactor: move stratification into stratify.rs
