+++ rhizome/src/logic/ast.rs
@@ -2,7 +2,6 @@ pub(super) mod body_term;
 pub(super) mod cid_value;
 pub(super) mod clause;
 pub(super) mod declaration;
-pub(super) mod dependency;
 pub(super) mod fact;
 pub(super) mod program;
 pub(super) mod rule;
@@ -13,7 +12,6 @@ pub(super) use body_term::*;
 pub(super) use cid_value::*;
 pub(super) use clause::*;
 pub(super) use declaration::*;
-pub(super) use dependency::*;
 pub(super) use fact::*;
 pub(super) use program::*;
 pub(super) use rule::*;
