+++ rhizome/src/logic/ast/clause.rs
@@ -1,10 +1,8 @@
-use derive_more::{From, IsVariant, TryInto};
-
 use crate::id::RelationId;
 
-use super::{Edge, Fact, Rule};
+use super::{Fact, Rule};
 
-#[derive(Debug, From, IsVariant, TryInto)]
+#[derive(Debug)]
 pub enum Clause {
     Fact(Fact),
     Rule(Rule),
@@ -15,13 +13,6 @@ impl Clause {
         match self {
             Clause::Fact(fact) => fact.head(),
             Clause::Rule(rule) => rule.head(),
-        }
-    }
-
-    pub fn depends_on(&self) -> Vec<Edge> {
-        match self {
-            Self::Fact(fact) => fact.depends_on(),
-            Self::Rule(rule) => rule.depends_on(),
         }
     }
 }
