+++ rhizome/src/logic/lower_to_ram.rs
@@ -5,11 +5,6 @@ use std::{
 };
 
 use anyhow::Result;
-use petgraph::{
-    graph::{DiGraph, NodeIndex},
-    visit::EdgeRef,
-    Direction,
-};
 
 use crate::{
     col_val::ColVal,
@@ -25,16 +20,12 @@ use crate::{
     value::Val,
 };
 
-use super::ast::{
-    cid_value::CidValue,
-    clause::Clause,
-    declaration::Declaration,
-    dependency::{Node, Polarity},
-    fact::Fact,
-    program::Program,
-    rule::Rule,
-    stratum::Stratum,
-    GetLink, Negation, RelPredicate, VarPredicate,
+use super::{
+    ast::{
+        cid_value::CidValue, declaration::Declaration, fact::Fact, program::Program, rule::Rule,
+        stratum::Stratum, GetLink, Negation, RelPredicate, VarPredicate,
+    },
+    stratify::stratify,
 };
 
 pub(crate) fn lower_to_ram<EF, IF, ER, IR>(
@@ -1069,91 +1060,4 @@ pub(crate) fn semi_naive_rewrites(rule: &Rule) -> Vec<Vec<(&RelPredicate, Relati
     }
 
     rewrites
-}
-
-pub(crate) fn stratify(program: &Program) -> Result<Vec<Stratum<'_>>> {
-    let mut clauses_by_relation = im::HashMap::<RelationId, im::Vector<&Clause>>::default();
-
-    for clause in program.clauses() {
-        clauses_by_relation = clauses_by_relation.alter(
-            |old| match old {
-                Some(clauses) => {
-                    let mut new = clauses;
-                    new.push_back(clause);
-
-                    Some(new)
-                }
-                None => Some(im::vector![clause]),
-            },
-            clause.head(),
-        );
-    }
-
-    let mut edg = DiGraph::<Node, Polarity>::default();
-    let mut nodes = im::HashMap::<Node, NodeIndex>::default();
-
-    for clause in program.clauses() {
-        nodes
-            .entry(Node::Idb(clause.head()))
-            .or_insert_with(|| edg.add_node(Node::Idb(clause.head())));
-
-        for dependency in clause.depends_on() {
-            nodes
-                .entry(dependency.to())
-                .or_insert_with(|| edg.add_node(dependency.to()));
-
-            nodes
-                .entry(dependency.from())
-                .or_insert_with(|| edg.add_node(dependency.from()));
-
-            let to = nodes
-                .get(&dependency.to())
-                .ok_or_else(|| Error::InternalRhizomeError("dependency not found".to_owned()))?;
-
-            let from = nodes
-                .get(&dependency.from())
-                .ok_or_else(|| Error::InternalRhizomeError("dependency not found".to_owned()))?;
-
-            edg.add_edge(*from, *to, dependency.polarity());
-        }
-    }
-
-    let sccs = petgraph::algo::kosaraju_scc(&edg);
-
-    for scc in &sccs {
-        for node in scc {
-            for edge in edg.edges_directed(*node, Direction::Outgoing) {
-                if edge.weight().is_negative() && scc.contains(&edge.target()) {
-                    return error(Error::ProgramUnstratifiable);
-                }
-            }
-        }
-    }
-
-    Ok(sccs
-        .iter()
-        .map(|nodes| {
-            let mut relations: HashSet<RelationId> = HashSet::default();
-            let mut clauses: Vec<&Clause> = Vec::default();
-
-            for i in nodes {
-                if let Some(Node::Idb(id)) = edg.node_weight(*i) {
-                    relations.insert(*id);
-
-                    if let Some(by_relation) = clauses_by_relation.get(id) {
-                        for clause in by_relation {
-                            clauses.push(*clause);
-                        }
-                    }
-                }
-            }
-
-            Stratum::new(
-                relations,
-                clauses,
-                nodes.len() > 1 || edg.contains_edge(nodes[0], nodes[0]),
-            )
-        })
-        .rev()
-        .collect())
 }
